+++ src/ppl/nn/engines/x86/impls/src/ppl/kernel/x86/fp32/roialign/roialign_n16cx_fp32_avx.cpp
@@ -51,7 +51,7 @@ ppl::common::RetCode roialign_n16cx_fp32_avx(
     const int64_t rois_offset = 4;
 
     PRAGMA_OMP_PARALLEL_FOR()
-    for (int i = 0; i < num_rois; i++) {
+    for (int64_t i = 0; i < num_rois; i++) {
         int64_t index_n             = i * pad_c * pooled_width * pooled_height;
         const float *p_rois         = rois + i * rois_offset;
         const int64_t roi_batch_ind = batch_indices[i];
